refactor: adopt component.tsx naming pattern with shorter directories
- Rename DesignSystemSection → DesignSystem
- Rename TechnicalImplementationSection → TechnicalImplementation
- Implement component.tsx pattern for all layout components
- Add index.ts files for clean imports
- Update all import paths to use shorter, cleaner references

Removes verbose directory names and repetitive file naming while
maintaining clear component organization.
