+++ src/domains/layout/DesignSystem/index.ts
@@ -1,0 +1,1 @@
+export { DesignSystemSection } from './component';
